# docs/swagger/swagger.yaml
info:
  contact: {}
  description: 超链 多币种钱包 -- Btc接口
  license: {}
  title: Bitcoin Wallet API
  version: "1.0"
paths:
  /btc/v1/account/balance/:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 比特币帐户地址
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: failed
          schema:
            type: string
        "500":
          description: failed
          schema:
            type: string
      summary: 获取余额
  /btc/v1/account/transactions/:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 比特币帐户地址
        in: query
        name: address
        required: true
        type: string
      - description: 起始页
        in: query
        name: offset
        type: string
      - description: 最多获取的记录总数
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: failed
          schema:
            type: string
        "500":
          description: failed
          schema:
            type: string
      summary: 获取交易列表
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 交易的签名后字符串
        in: formData
        name: tx
        required: true
        type: string
      - description: 发送方地址
        in: formData
        name: from
        required: true
        type: string
      - description: 接收方地址
        in: formData
        name: to
        required: true
        type: string
      - description: 矿工费
        in: formData
        name: fee
        required: true
        type: string
      - description: 交易额
        in: formData
        name: amount
        required: true
        type: string
      - description: 备注
        in: formData
        name: remark
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: failed
          schema:
            type: string
        "500":
          description: failed
          schema:
            type: string
      summary: 发送交易
  /btc/v1/account/unspents/:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 比特币帐户地址
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: failed
          schema:
            type: string
        "500":
          description: failed
          schema:
            type: string
      summary: 获取未花列表
  /btc/v1/wallet/accounts/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 钱包的Id
        in: formData
        name: walletid
        required: true
        type: integer
      - description: BTC地址
        in: formData
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: 添加BTC帐户成功
          schema:
            type: string
        "400":
          description: 非法参数
          schema:
            type: string
        "500":
          description: 失败
          schema:
            type: string
      summary: 向钱包增加BTC帐户（因产品设计成比特币与以太币分开为单独的钱包，故该API用不上）
  /btc/v1/wallet/phone/:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 钱包的Id
        in: formData
        name: walletid
        required: true
        type: integer
      - description: 电话号码
        in: formData
        name: phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 修改电话号码成功
          schema:
            type: string
        "400":
          description: 非法参数
          schema:
            type: string
        "500":
          description: 失败
          schema:
            type: string
      summary: 修改钱包的电话号码
  /btc/v1/wallets/:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 钱包的Id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 删除钱包成功
          schema:
            type: string
        "400":
          description: 非法参数
          schema:
            type: string
        "500":
          description: 失败
          schema:
            type: string
      summary: 根据指定的钱包Id删除钱包
swagger: "2.0"